Moved pure-Python dx7 simulator to its own file.
